requirements.txt
websockets
mega.py
asyncio
python-dotenv